--- FormBarOauth/package.json
@@ -1,9 +1,0 @@
-{
-  "dependencies": {
-    "express": "^5.1.0",
-    "express-session": "^1.18.2",
-    "fetch": "^1.1.0",
-    "jsonwebtoken": "^9.0.2",
-    "node-fetch": "^3.3.2"
-  }
-}
